Split the AutoResponseHandler into each handler
